Allow the compiler to use zstd_marshal
When zstd is available, build bytecode runtime executables with zstd
support. This enables it for the entire build. The .opt tools are of
course compiled with a full compiler, so can use marshal_zstd.cmxa in
the "normal" way.
