+++ .gitignore
@@ -221,6 +221,7 @@ META
 /otherlibs/dynlink/dynlink_compilerlibs/*.ml
 /otherlibs/dynlink/dynlink_compilerlibs/*.mli
 /otherlibs/dynlink/dynlink_compilerlibs/.depend
+/otherlibs/marshal_zstd/zstd.c
 /otherlibs/unix/unix.ml
 
 /parsing/parser.ml
